pyproject.toml
[project]
name = "av_meta"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Flet developer", email = "you@example.com" }
]
dependencies = [
    "blinker>=1.9.0",
    "flet>=0.80.5",
    "tortoise-orm>=1.1.5",
    "xmltodict>=1.0.3",
]

[dependency-groups]
dev = [
    "flet-cli>=0.80.5",
    "flet-desktop>=0.80.5",
    "flet-web>=0.80.5",
]

[tool.flet]
org = "com.mycompany"
product = "av_meta"
company = "Flet"
copyright = "Copyright (C) 2023-2026 by Flet"

[tool.flet.app]
path = "src"
